Have made some more tests based off the suggestions from the tutorial.
This involves making tests for the results view as well as modifying
and adding to some of the older tests so that no question will appear
on the index page unless it has a choice. Only one right now to pass.
Also no detail or result page should load if the question has no
choice.